Simplify jwk locator to use in various oidc use cases
